requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.7.4
click==8.1.7
databases==0.9.0
dill==0.3.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.5
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isort==5.13.2
Jinja2==3.1.4
libcst==1.4.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
pep8==1.7.1
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylint==3.2.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlmodel==0.0.21
starlette==0.37.2
strawberry-graphql==0.237.3
tomlkit==0.13.0
typer==0.12.3
typing_extensions==4.12.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
